refactor: おてんきぐらしの責務を分割
