+++ src/app/otenkigurashi/otenkigurashiCopy.ts
@@ -1,0 +1,60 @@
+export const OTENKI_COPY = {
+        ja: {
+            react1: 'えへへ！',
+            react2: 'おさんぽ行く？',
+            react3: 'ふわああ…',
+            react4: 'なになに？',
+            react5: '今日もいいお天気！',
+            lead: '天気予報を見ないあなたの、',
+            lead2: 'いちばん優しいおまもり。',
+            lead3: 'おてんきぐらし',
+            lead4: 'は、',
+            lead5: '現実の天気と連動するシミュレーションです。',
+            conceptTitle: 'コンセプト',
+            conceptClick: 'ゲームみたいに天気を楽しめるんだ！',
+            conceptText: '天気予報の確認は面倒だけど、急な雨や気圧の変化はつらい… そんな方々のために生まれました。ゲーム性のある優しい世界を通して、面倒だった天気確認を「雨だから、ゲーム内で特別なことができるかも？」という、ポジティブな体験へと変えていきます。',
+            f1Click: '大阪で雨なら、ゲームでも雨だよ！',
+            f1Title: '☀ リアルタイム天気連動',
+            f1Text: 'あなたのいる場所の「今」が、キャラクターの世界に直接反映されます。雨は雨粒、雷はフラッシュ、雪は降雪モーションとして画面に現れ、夜になればキャラクターも眠りにつきます。',
+            f2Click: 'おさんぽで色んなアイテムを集めよう！',
+            f2Title: '🚶 おさんぽ (Walking)',
+            f2Text: 'キャラクターを「おさんぽ」に出すことができます。おさんぽ先の景色や、手に入るアイテムは天気によって変化。コレクションする楽しみが待っています。',
+            f3Click: '思い出がたくさん記録されるよ！',
+            f3Title: '✨ コレクション＆実績',
+            f3Text: 'レベルアップのようなノルマはありません。「おさんぽ」で集めたアイテムを眺める「ずかん」や、「はじめて雨の日におさんぽした」といったキャラクターとの思い出を記録する「実績」が、あなたの毎日を彩ります。',
+            techClick: 'Next.jsで作られてるよ！',
+            openApp: 'アプリをひらく',
+            viewGithub: 'GitHubをみる',
+            backHome: 'おうちにもどる',
+            logoAlt: 'OTENKI GURASHI Logo',
+        },
+        en: {
+            react1: 'Hehe!',
+            react2: 'Want to go for a walk?',
+            react3: 'Yaaawn...',
+            react4: 'What is it?',
+            react5: 'Nice weather today too!',
+            lead: 'For you who never checks the forecast,',
+            lead2: 'the kindest little charm.',
+            lead3: 'Otenkigurashi',
+            lead4: 'is',
+            lead5: 'a simulation linked to real-world weather.',
+            conceptTitle: 'CONCEPT',
+            conceptClick: 'You can enjoy weather like a game!',
+            conceptText: 'Checking weather forecasts can be a hassle, and sudden rain or pressure changes can be tough. This app was born for that. Through a gentle game-like world, it turns weather checks from a chore into a positive experience: “It is raining, maybe I can do something special in-game.”',
+            f1Click: 'If it rains in Osaka, it rains in the game too!',
+            f1Title: '☀ Real-Time Weather Sync',
+            f1Text: 'The “now” of your location is directly reflected in the character world. Rain appears as droplets, thunder as flashes, snow as falling motion, and at night the character goes to sleep.',
+            f2Click: 'Let us collect items on a walk!',
+            f2Title: '🚶 Walk Mode',
+            f2Text: 'You can send your character on a walk. Scenery and obtainable items change with the weather, giving you a fun collection loop.',
+            f3Click: 'So many memories get recorded!',
+            f3Title: '✨ Collection & Achievements',
+            f3Text: 'There are no level-up quotas. A “zukan” lets you look back at collected walk items, and achievements record memories with your character, like your first rainy-day walk.',
+            techClick: 'It is built with Next.js!',
+            openApp: 'Open App',
+            viewGithub: 'View GitHub',
+            backHome: 'Back Home',
+            logoAlt: 'OTENKIGURASHI Logo',
+        },
+    } as const;
